IMP - Move `getImageURL` to unit info utils
